requirements.txt
certifi==2019.11.28
Click==7.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
pandas==0.25.3
python-dateutil==2.8.1
pickle
base64
io
pytz==2019.3
scikit-learn
scipy
six
Werkzeug
wincertstore
Flask==1.1.2
Flask_Cors==3.0.10
numpy==1.20.1
pandas==1.2.4
Pillow==8.4.0
tensorflow==2.3.0
